AppStateProvider handles fetching language data
