+++ frontend/app_v2/src/bootstrap.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
+import { Switch, Route } from 'react-router-dom'
 import AppV1Provider from 'app_v1/FVProvider'
 import 'tailwindcss/tailwind.css'
 import { BrowserRouter as Router } from 'react-router-dom'
@@ -10,13 +11,17 @@ const queryClient = new QueryClient()
 
 ReactDOM.render(
   <AppV1Provider>
-    <Router>
-      <AppStateProvider>
-        <QueryClientProvider client={queryClient}>
-          <AppFrameContainer />
-        </QueryClientProvider>
-      </AppStateProvider>
-    </Router>
+    <QueryClientProvider client={queryClient}>
+      <Router>
+        <Switch>
+          <Route path="/:language">
+            <AppStateProvider>
+              <AppFrameContainer />
+            </AppStateProvider>
+          </Route>
+        </Switch>
+      </Router>
+    </QueryClientProvider>
   </AppV1Provider>,
   document.getElementById('root')
 )
